runner: reinitialize locale in worker.
The runner calls setlocale(LC_ALL, "") to allow unicode-enabled
functions to work. Previously, we did not have problems with that
since fork() kept that intialization in the new process.

This was no longer the case after BoxFort was introduced. We need to
call setlocale again to make, for instance, "%ls" work with cr_log_*
functions.

Addresses #216.